anyupdate/requirement.txt
backports.csv==1.0.7
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.1
django-dashing==0.3.4
django-grappelli==2.13.1
django-import-export==1.2.0
et-xmlfile==1.0.1
httpretty==0.9.6
idna==2.8
ijson==2.4
jdcal==1.4.1
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==2.6.3
pysnow==0.7.9
pytz==2019.1
#PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
SOAPpy-py3==0.52.24
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.25.3
wincertstore==0.2
wstools-py3==0.54.4
xlrd==1.2.0
xlwt==1.3.0
djangorestframework==3.10.2
